requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
joblib==1.0.1
reportlab==3.6.1
python-dotenv==0.19.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1
